# Conflicts:
#	src/main/java/com/app/kokonut/apiKey/ApiKeyRestController.java
#	src/main/java/com/app/kokonut/apiKey/ApiKeyService.java
#	src/main/java/com/app/kokonut/collectInformation/CollectInformationRestController.java
#	src/main/java/com/app/kokonut/email/email/EmailRestController.java
#	src/main/java/com/app/kokonut/email/emailGroup/EmailGroupRestController.java
#	src/main/java/com/app/kokonut/email/emailGroup/EmailGroupService.java
#	src/main/java/com/app/kokonut/faq/FaqRestController.java
#	src/main/java/com/app/kokonut/faq/FaqService.java
#	src/main/java/com/app/kokonut/notice/NoticeRestController.java
#	src/main/java/com/app/kokonut/notice/NoticeService.java
#	src/main/java/com/app/kokonut/qna/QnaRestController.java
#	src/main/java/com/app/kokonut/qna/QnaService.java
#	src/main/java/com/app/kokonut/service/KnServiceService.java
#	src/main/java/com/app/kokonut/setting/KnSettingRestController.java
#	src/main/java/com/app/kokonut/setting/SettingService.java

+++ src/main/java/com/app/kokonut/bizMessage/friendtalkMessage/FriendtalkMessageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FriendtalkMessageRepository extends JpaRepository<FriendtalkMessage, Integer>, JpaSpecificationExecutor<FriendtalkMessage>, FriendtalkMessageRepositoryCustom {
+public interface FriendtalkMessageRepository extends JpaRepository<FriendtalkMessage, Long>, JpaSpecificationExecutor<FriendtalkMessage>, FriendtalkMessageRepositoryCustom {
 
 }
