+++ src/main/java/com/app/kokonut/bizMessage/friendtalkMessage/friendtalkMessageRecipient/FriendtalkMessageRecipientRepository.java
@@ -3,6 +3,6 @@ package com.app.kokonut.bizMessage.friendtalkMessage.friendtalkMessageRecipient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FriendtalkMessageRecipientRepository extends JpaRepository<FriendtalkMessageRecipient, Integer>, JpaSpecificationExecutor<FriendtalkMessageRecipient> {
+public interface FriendtalkMessageRecipientRepository extends JpaRepository<FriendtalkMessageRecipient, Long>, JpaSpecificationExecutor<FriendtalkMessageRecipient> {
 
 }
