+++ src/main/java/com/app/kokonut/email/emailHistory/EmailHistoryRepository.java
@@ -1,9 +1,8 @@
 package com.app.kokonut.email.emailHistory;
 
-import com.app.kokonut.email.emailHistory.entity.EmailHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface EmailHistoryRepository extends JpaRepository<EmailHistory, Integer>, JpaSpecificationExecutor<EmailHistory> {
+public interface EmailHistoryRepository extends JpaRepository<EmailHistory, Long>, JpaSpecificationExecutor<EmailHistory> {
 
 }
