+++ src/main/java/com/app/kokonut/faq/FaqRepository.java
@@ -1,9 +1,8 @@
 package com.app.kokonut.faq;
 
-import com.app.kokonut.faq.entity.Faq;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FaqRepository extends JpaRepository<Faq, Integer>, JpaSpecificationExecutor<Faq>, FaqRepositoryCustom {
+public interface FaqRepository extends JpaRepository<Faq, Long>, JpaSpecificationExecutor<Faq>, FaqRepositoryCustom {
 
 }
