+++ src/main/java/com/app/kokonut/qna/QnaRepository.java
@@ -1,9 +1,8 @@
 package com.app.kokonut.qna;
 
-import com.app.kokonut.qna.entity.Qna;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface QnaRepository extends JpaRepository<Qna, Integer>, JpaSpecificationExecutor<Qna>, QnaRepositoryCustom {
+public interface QnaRepository extends JpaRepository<Qna, Long>, JpaSpecificationExecutor<Qna>, QnaRepositoryCustom {
 
 }
